refactor(api/plant): split generateNode() into plantService
TodoService.complete()にPlantの成長ロジックが混在していたため、
ノード生成の責務をPlantServiceへ分離する。依存するrandom()ヘルパーと
型定義(CreatedNode, PlantWithNode)も合わせてplant.types.tsへ移動。
