+++ apps/api/src/todo/todo.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { TodoService } from './todo.service';
 import { TodoController } from './todo.controller';
 import { GardenModule } from 'src/garden/garden.module';
+import { PlantModule } from 'src/plant/plant.module';
 
 @Module({
-  imports: [GardenModule],
+  imports: [GardenModule, PlantModule],
   controllers: [TodoController],
   providers: [TodoService],
 })
