+++ apps/api/src/todo/types/todo.types.ts
@@ -1,18 +1,8 @@
-import { GrowthStage, Prisma, TodoScore } from 'generated/prisma/client';
+import { GrowthStage, TodoScore } from 'generated/prisma/client';
 
 export type Score = {
   rank?: TodoScore;
   nodeCount: number;
-};
-
-export type CreatedNode = {
-  hue: number;
-  size: number;
-  length: number;
-  depth: number;
-  parentId: string;
-  todoId: string;
-  plantId: string;
 };
 
 export type CompleteTodo = {
@@ -20,10 +10,6 @@ export type CompleteTodo = {
   completedAt: Date;
   targetDuration: number | null;
 };
-
-export type PlantWithNode = Prisma.PlantGetPayload<{
-  include: { plantNodes: true };
-}>;
 
 export type GrowthStageResult = {
   curStage: GrowthStage;
